pyproject.toml
[project]
name = "ollama-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "openai-agents>=0.0.19",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]
